+++ lib/egis/output_downloader.rb
@@ -5,7 +5,7 @@ require 'csv'
 module Egis
   # @!visibility private
   class OutputDownloader
-    def initialize(aws_s3_client:)
+    def initialize(aws_s3_client)
       @s3_client = aws_s3_client
     end