+++ lib/egis/s3_cleaner.rb
@@ -3,7 +3,7 @@
 module Egis
   # @!visibility private
   class S3Cleaner
-    def initialize(aws_s3_client:)
+    def initialize(aws_s3_client)
       @s3_client = aws_s3_client
     end